Fix bugs server

+++ server/routes/activities.js
@@ -5,10 +5,13 @@ const router = express.Router();
 const { isAdmin, isLoggedIn } = require("../middleware/account-middleware");
 
 const Activity = require("../models/Activity");
+const Class = require("../models/Class");
 
 router.get("/", async (req, res, next) => {
   try {
-    return res.status(200).json(await Activity.find());
+    return res
+      .status(200)
+      .json(await Activity.find().select("-updatedAt -createdAt -__v"));
   } catch (e) {
     return next(e);
   }
@@ -24,16 +27,18 @@ router.post("/create", isLoggedIn(), isAdmin(), async (req, res, next) => {
         .json(await Activity.create({ name, description, type }));
     else
       return res.status(401).json({
-        status: "ActivityExists"
+        status: "ActivityExists",
       });
   } catch (e) {
-    return next(e);
+    return e.name === "ValidationError"
+      ? res.status(400).json({ status: "ValidationError", error: e.errors })
+      : next(e);
   }
 });
 
 router.get("/:id", async (req, res, next) => {
   try {
-    const activity = await Activity.findOne({ _id: req.params.id }).select(
+    const activity = await Activity.findById({ _id: req.params.id }).select(
       "-updatedAt -createdAt -__v"
     );
 
@@ -41,7 +46,9 @@ router.get("/:id", async (req, res, next) => {
       ? res.status(200).json(activity)
       : res.status(400).json({ status: "BadRequest" });
   } catch (e) {
-    return next(e);
+    return e.name === "CastError"
+      ? res.status(400).json({ status: "BadRequest" })
+      : next(e);
   }
 });
 
@@ -49,33 +56,48 @@ router.put("/:id", isLoggedIn(), isAdmin(), async (req, res, next) => {
   try {
     const { activity } = req.body;
 
-    const updatedActivity = await Activity.findByIdAndUpdate(
-      { _id: req.params.id },
-      {
-        name: activity.name,
-        desciption: activity.desciption,
-        type: activity.type
-      },
-      { new: true, runValidators: true }
-    );
+    if (!(await Activity.findOne({ name: activity.name })))
+      if ((await Class.find({ activity: req.params.id })).length === 0) {
+        const updatedActivity = await Activity.findByIdAndUpdate(
+          { _id: req.params.id },
+          {
+            name: activity.name,
+            desciption: activity.desciption,
+            type: activity.type,
+          },
+          { new: true, runValidators: true }
+        );
 
-    return updatedActivity
-      ? res.status(200).json(updatedActivity)
-      : res.status(400).json({ status: "BadRequest" });
+        return updatedActivity
+          ? res.status(200).json(updatedActivity)
+          : res.status(400).json({ status: "BadRequest" });
+      } else return res.status(406).json({ ststus: "NotAcceptable" });
+    else
+      return res.status(401).json({
+        status: "ActivityExists",
+      });
   } catch (e) {
-    return next(e);
+    return e.name === "CastError"
+      ? res.status(400).json({ status: "BadRequest" })
+      : e.name === "ValidationError"
+      ? res.status(400).json({ status: "ValidationError", error: e.errors })
+      : next(e);
   }
 });
 
 router.delete("/:id", isLoggedIn(), isAdmin(), async (req, res, next) => {
   try {
-    const result = await Activity.deleteOne({ _id: req.params.id });
+    if ((await Class.find({ activity: req.params.id })).length === 0) {
+      const result = await Activity.deleteOne({ _id: req.params.id });
 
-    return result.n
-      ? res.status(200).json({ status: "OperationSuccessful" })
-      : res.status(400).json({ status: "BadRequest" });
+      return result.n
+        ? res.status(200).json({ status: "OperationSuccessful" })
+        : res.status(400).json({ status: "BadRequest" });
+    } else return res.status(406).json({ status: "NotAcceptable" });
   } catch (e) {
-    return next(e);
+    return e.name === "CastError"
+      ? res.status(400).json({ status: "BadRequest" })
+      : next(e);
   }
 });
